update platform browser

+++ FrontEnd/src/app/service/session-storage.service.ts
@@ -1,22 +1,26 @@
-import { Injectable } from '@angular/core';
+import { Injectable, Inject, PLATFORM_ID } from '@angular/core';
+import { isPlatformBrowser } from '@angular/common';
 
 @Injectable({
   providedIn: 'root'
 })
 export class SessionStorageService {
 
-  constructor() { }
+  constructor(@Inject(PLATFORM_ID) private platformId: Object) { }
 
   setItem(key: string, value: any, expiryInMinutes: number): void {
+    if (isPlatformBrowser(this.platformId)) {
     const now = new Date();
     const item = {
       value: value,
       expiry: now.getTime() + expiryInMinutes * 60 * 1000, // Convert minutes to milliseconds
     };
     sessionStorage.setItem(key, JSON.stringify(item));
+    }
   }
 
   getItem(key: string): any {
+    if (isPlatformBrowser(this.platformId)) {
     const itemStr = sessionStorage.getItem(key);
     if (!itemStr) {
       return null;
@@ -31,13 +35,18 @@ export class SessionStorageService {
       return null;
     }
     return item.value;
+    }
   }
 
   removeItem(key: string): void {
-    sessionStorage.removeItem(key);
+    if (isPlatformBrowser(this.platformId)) {
+      sessionStorage.removeItem(key);
+    }
   }
 
   clear(): void {
-    sessionStorage.clear();
+    if (isPlatformBrowser(this.platformId)) {
+      sessionStorage.clear();
+    }
   }
 }
